# api-docs.json
{"openapi":"3.0.1","info":{"title":"GlobeTales API","description":"API documentation for the GlobeTales project","version":"1.0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"User","description":"User management APIs"}],"paths":{"/api/users/{id}":{"get":{"tags":["User"],"summary":"Get user by ID","description":"Returns a single user by their UUID.","operationId":"getUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"put":{"tags":["User"],"summary":"Update user","description":"Updates the details of an existing user by UUID.","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}},"delete":{"tags":["User"],"summary":"Delete user","description":"Deletes a user by their UUID.","operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/users/{id}/location":{"get":{"tags":["User"],"summary":"Get user location","description":"Returns the location of a user by UUID.","operationId":"getUserLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocationDTO"}}}}}},"put":{"tags":["User"],"summary":"Update user location","description":"Updates the location of a user by UUID.","operationId":"updateUserLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"latitude","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"longitude","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK"}}},"post":{"tags":["User"],"summary":"Set user location","description":"Sets the location of a user by UUID.","operationId":"setUserLocation","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/users/location":{"put":{"tags":["User"],"summary":"Update current user location","description":"Updates the location of the currently authenticated user.","operationId":"updateCurrentUserLocation","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDTO"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/api/stories/{id}":{"get":{"tags":["user-story-controller"],"operationId":"getStoryById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserStoryDTO"}}}}}},"put":{"tags":["user-story-controller"],"operationId":"updateStory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStoryDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserStoryDTO"}}}}}},"delete":{"tags":["user-story-controller"],"operationId":"deleteStory","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/reviews/{id}":{"put":{"tags":["place-review-controller"],"operationId":"updateReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceReviewDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceReviewDTO"}}}}}},"delete":{"tags":["place-review-controller"],"operationId":"deleteReview","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/api/visited-sites":{"post":{"tags":["visited-site-controller"],"operationId":"addVisitedSite","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VisitedSite"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/VisitedSite"}}}}}}},"/api/users":{"get":{"tags":["User"],"summary":"Get all users","description":"Returns a list of all users.","operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"post":{"tags":["User"],"summary":"Create a new user","description":"Creates a new user with the provided details and password.","operationId":"createUser","parameters":[{"name":"password","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/user-favourites":{"post":{"tags":["user-favourite-controller"],"operationId":"addFavourite","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserFavouriteDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserFavouriteDTO"}}}}}}},"/api/stories":{"get":{"tags":["user-story-controller"],"operationId":"getStoriesForUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserStoryDTO"}}}}}}},"post":{"tags":["user-story-controller"],"operationId":"createStory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserStoryDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserStoryDTO"}}}}}}},"/api/stories/visit/{placeId}":{"post":{"tags":["user-story-controller"],"operationId":"markSiteAsVisited","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserStoryDTO"}}}}}}},"/api/routes":{"post":{"tags":["route-controller"],"operationId":"getRoute","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RouteRequestDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RouteResponseDTO"}}}}}}},"/api/reviews":{"post":{"tags":["place-review-controller"],"operationId":"createReview","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PlaceReviewDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PlaceReviewDTO"}}}}}}},"/api/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/places/update-area":{"post":{"tags":["place-controller"],"operationId":"updateSitesForArea","parameters":[{"name":"bbox","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/places/fetch":{"post":{"tags":["place-controller"],"operationId":"fetchPlacesForArea","parameters":[{"name":"bbox","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/chat":{"post":{"tags":["chatbot-controller"],"operationId":"chat","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatbotRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatbotResponse"}}}}}}},"/api/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthenticationRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthenticationResponse"}}}}}}},"/api/weather":{"get":{"tags":["weather-controller"],"operationId":"getWeather","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WeatherResponseDTO"}}}}}}},"/api/visited-sites/user/{userId}":{"get":{"tags":["visited-site-controller"],"operationId":"getVisitedSitesByUser","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VisitedSite"}}}}}}}},"/api/visited-sites/check":{"get":{"tags":["visited-site-controller"],"operationId":"isSiteVisited","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string"}},{"name":"placeId","in":"query","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/users/me":{"get":{"tags":["User"],"summary":"Get current user","description":"Returns the currently authenticated user.","operationId":"getCurrentUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDTO"}}}}}}},"/api/users/check/username":{"get":{"tags":["User"],"summary":"Check if username exists","description":"Checks if a username is already registered.","operationId":"checkUsernameExists","parameters":[{"name":"username","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/users/check/email":{"get":{"tags":["User"],"summary":"Check if email exists","description":"Checks if an email is already registered.","operationId":"checkEmailExists","parameters":[{"name":"email","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"boolean"}}}}}}},"/api/user-favourites/user/{userId}":{"get":{"tags":["user-favourite-controller"],"operationId":"getFavouritesByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserFavouriteDTO"}}}}}}}},"/api/user-favourites/user/{userId}/type/{siteType}":{"get":{"tags":["user-favourite-controller"],"operationId":"getFavouritesByType","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"siteType","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserFavouriteDTO"}}}}}}}},"/api/user-achievements/user/{userId}":{"get":{"tags":["user-achievement-controller"],"operationId":"getAchievementsByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserAchievementDTO"}}}}}}}},"/api/stories/visited":{"get":{"tags":["user-story-controller"],"operationId":"getVisitedSitesForUser","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserStoryDTO"}}}}}}}},"/api/reviews/user/{userId}":{"get":{"tags":["place-review-controller"],"operationId":"getReviewsByUserId","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaceReviewDTO"}}}}}}}},"/api/reviews/place/{placeId}":{"get":{"tags":["place-review-controller"],"operationId":"getReviewsByPlaceId","parameters":[{"name":"placeId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PlaceReviewDTO"}}}}}}}},"/api/recommendations":{"get":{"tags":["recommendation-controller"],"operationId":"getRecommendations","parameters":[{"name":"transportMode","in":"query","required":true,"schema":{"type":"string"}},{"name":"maxDistance","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":0}},{"name":"size","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":10}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RecommendationResponseDTO"}}}}}}},"/api/places":{"get":{"tags":["place-controller"],"operationId":"getAllPlaces","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Place"}}}}}}}},"/api/places/{id}":{"get":{"tags":["place-controller"],"operationId":"getPlaceById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Place"}}}}}}},"/api/chat/greeting":{"get":{"tags":["chatbot-controller"],"operationId":"getGreeting","parameters":[{"name":"lat","in":"query","required":true,"schema":{"type":"number","format":"double"}},{"name":"lon","in":"query","required":true,"schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatbotResponse"}}}}}}},"/api/visited-sites/user/{userId}/place/{placeId}":{"delete":{"tags":["visited-site-controller"],"operationId":"removeVisitedSite","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string"}},{"name":"placeId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/api/user-favourites/user/{userId}/site/{siteId}":{"delete":{"tags":["user-favourite-controller"],"operationId":"deleteFavouriteByUserAndSite","parameters":[{"name":"userId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"siteId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"UserDTO":{"required":["displayName","email","username"],"type":"object","properties":{"id":{"type":"string","format":"uuid"},"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"type":"string"},"displayName":{"maxLength":100,"minLength":2,"type":"string"},"roles":{"uniqueItems":true,"type":"array","items":{"type":"string"}},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"experiencePoints":{"type":"integer","format":"int32"},"level":{"type":"string"}}},"LocationDTO":{"type":"object","properties":{"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"UserStoryDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"placeId":{"type":"string"},"title":{"type":"string"},"content":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"visitDate":{"type":"string","format":"date-time"}}},"PlaceReviewDTO":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"username":{"type":"string"},"placeId":{"type":"string"},"placeName":{"type":"string"},"comment":{"type":"string"},"rating":{"type":"integer","format":"int32"},"createdAt":{"type":"string","format":"date-time"}}},"VisitedSite":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"string"},"placeId":{"type":"integer","format":"int64"},"placeName":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"visitedAt":{"type":"string","format":"date-time"}}},"UserFavouriteDTO":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"siteId":{"type":"integer","format":"int64"},"siteType":{"type":"string"},"savedAt":{"type":"string","format":"date-time"},"placeName":{"type":"string"}}},"PointDTO":{"type":"object","properties":{"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"RouteRequestDTO":{"type":"object","properties":{"start":{"$ref":"#/components/schemas/PointDTO"},"end":{"$ref":"#/components/schemas/PointDTO"},"transportMode":{"type":"string"}}},"Geometry":{"type":"object","properties":{"type":{"type":"string"},"coordinates":{"type":"array","items":{"type":"array","items":{"type":"number","format":"double"}}}}},"RouteResponseDTO":{"type":"object","properties":{"distance":{"type":"number","format":"double"},"duration":{"type":"number","format":"double"},"geometry":{"$ref":"#/components/schemas/Geometry"}}},"RegisterRequest":{"required":["email","password","username"],"type":"object","properties":{"username":{"maxLength":50,"minLength":3,"type":"string"},"email":{"type":"string"},"password":{"maxLength":2147483647,"minLength":6,"type":"string"}}},"AuthenticationResponse":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDTO"}}},"ChatbotRequest":{"type":"object","properties":{"message":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"}}},"ChatbotResponse":{"type":"object","properties":{"reply":{"type":"string"}}},"AuthenticationRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"CurrentWeatherDTO":{"type":"object","properties":{"temperature_2m":{"type":"number","format":"double"},"weather_code":{"type":"integer","format":"int32"}}},"WeatherResponseDTO":{"type":"object","properties":{"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"current":{"$ref":"#/components/schemas/CurrentWeatherDTO"}}},"UserAchievementDTO":{"type":"object","properties":{"achievementCode":{"type":"string"},"achievedAt":{"type":"string","format":"date-time"},"displayName":{"type":"string"},"description":{"type":"string"}}},"Place":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"osmId":{"type":"integer","format":"int64"},"name":{"type":"string"},"type":{"type":"string"},"latitude":{"type":"number","format":"double"},"longitude":{"type":"number","format":"double"},"address":{"type":"string"},"tags":{"type":"object","additionalProperties":{"type":"object"}}}},"RecommendationDTO":{"type":"object","properties":{"place":{"$ref":"#/components/schemas/Place"},"distance":{"type":"number","format":"double"},"duration":{"type":"number","format":"double"}}},"RecommendationResponseDTO":{"type":"object","properties":{"recommendations":{"type":"array","items":{"$ref":"#/components/schemas/RecommendationDTO"}},"hasMore":{"type":"boolean"}}}}}}